Detected Overfitting

+++ Exercises/Wie viel Ihr Auto Wert ist/Projekt_Woche8/RandomForest.py
@@ -13,7 +13,7 @@ x_train, x_test, y_train, y_test = train_test_split(x, y, test_size=0.2, random_
 MODEL_FILE_PATH = os.path.join(system_paths[0], "Model_Data", 'random_forest_model.pkl')
 
 
-if __name__ == "__main__" and os.path.exists(MODEL_FILE_PATH):
+if False and __name__ == "__main__" and os.path.exists(MODEL_FILE_PATH):
     os.remove(MODEL_FILE_PATH)
 
 if os.path.exists(MODEL_FILE_PATH):
@@ -45,17 +45,20 @@ else:
     
     joblib.dump(model, MODEL_FILE_PATH)
 
+y_pred = y_pred = model.predict(x_test)
+feature_importance_df = pd.DataFrame({'Feature': x.columns, 'Importance': model.feature_importances_})
+feature_importance_df = feature_importance_df.sort_values(by='Importance', ascending=False)
 
-# Metriken ausgeben
 if __name__ == '__main__':
     mse = mean_squared_error(y_test, y_pred)
-    print(f"Beste Hyperparameter: {best_params}")
     print(f"MSE: {mse:.4f}")
     print(f"RMSE: {mse**0.5:.4f}")
 
-    # Feature Importances anzeigen
-    importances = best_model.feature_importances_
-    feature_importance_df = pd.DataFrame({'Feature': x.columns, 'Importance': importances})
-    # Sortieren nach Importance in absteigender Reihenfolge
-    feature_importance_df = feature_importance_df.sort_values(by='Importance', ascending=False)
+    
+    y_pred = model.predict(x_train)
+    mse = mean_squared_error(y_train, y_pred)
+
+    print(f"MSE(Train): {mse:.4f}")
+    print(f"RMSE(Train): {mse**0.5:.4f}")
+    
     print(feature_importance_df)
